pyproject.toml
[tool.poetry]
name = "blazebase"
version = "0.1.0"
description = ""
authors = ["Charles Dupont <charlo.dupont@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
firebase-admin = "^6.4.0"
requests = "^2.31.0"
gcloud = "^0.18.3"
google-auth = "^2.27.0"
requests-toolbelt = "^1.0.0"
jwcrypto = "^1.5.1"
cryptography = "^42.0.1"
google-auth-oauthlib = "^1.2.0"
google-auth-httplib2 = "^0.2.0"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
